# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智能墨水屏平台 API
  version: 1.0.0
  description: |
    用户端/管理端/设备端统一接口文档。
    设备（真机与模拟）推荐统一启动链路：login -> auto-register -> bind/status -> login -> stream。
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/tf/{fileId}/download:
    get:
      summary: 下载TF卡文件
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: file stream
...